Encapsulate
